Usage of long refactored to int in LinkedHashMapRemoveEldest class

+++ application/src/main/java/org/thingsboard/server/actors/ActorSystemContext.java
@@ -489,7 +489,7 @@ public class ActorSystemContext {
 
     @Value("${actors.session.max_concurrent_sessions_per_device:1}")
     @Getter
-    private long maxConcurrentSessionsPerDevice;
+    private int maxConcurrentSessionsPerDevice;
 
     @Value("${actors.session.sync.timeout:10000}")
     @Getter
